+++ changelog.md
@@ -1,6 +1,11 @@
 # Retool changelog
 
 
+## 2.3.8 (2024-05-06)
+
+-   **_Fix_**: Fixed post filters not working.
+
+
 ## 2.3.7 (2024-04-28)
 
 -   **_Fix_**: Fixed clone list and `internal-config.json` minimum version detection.